# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexpoint Funnel IQ — AI-assisted funnel drop-off diagnosis tool and Happy Mom campaign tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FunnelEvent {
  id: number;
  eventType: string;
  sessionId: string;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export type EventInputEventType =
  (typeof EventInputEventType)[keyof typeof EventInputEventType];

export const EventInputEventType = {
  page_view: "page_view",
  banner_click: "banner_click",
  product_view: "product_view",
  add_to_cart: "add_to_cart",
  add_to_wishlist: "add_to_wishlist",
  checkout_start: "checkout_start",
  purchase: "purchase",
  subscription_start: "subscription_start",
  subscription_cancel: "subscription_cancel",
  nappy_subscription_click: "nappy_subscription_click",
  sale_item_view: "sale_item_view",
  browse_only: "browse_only",
} as const;

export interface EventInput {
  eventType: EventInputEventType;
  sessionId: string;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  metadata?: string | null;
}

export interface FunnelSummary {
  totalVisitors: number;
  bannerClicks: number;
  bannerCTR: number;
  productViews: number;
  saleItemViews?: number;
  browseOnlyCount?: number;
  addToCart: number;
  addToWishlist: number;
  checkoutStarts: number;
  purchases: number;
  conversionRate: number;
  repeatCustomerRate: number;
}

export interface DailyRevenue {
  date: string;
  revenue: number;
  orders: number;
}

export interface CampaignMetrics {
  campaignName: string;
  bannerClicks: number;
  discountItemViews: number;
  discountItemPurchases: number;
  browseOnlyVisitors: number;
  nappySubscriptions: number;
  totalRevenue: number;
  revenueByDay?: DailyRevenue[];
}

export interface TrafficPoint {
  label: string;
  visitors: number;
}

export interface TrafficData {
  activeNow: number;
  totalToday: number;
  totalThisWeek: number;
  hourlyTraffic: TrafficPoint[];
  dailyTraffic: TrafficPoint[];
}

export interface MonthlyCustomerPoint {
  month: string;
  newCustomers: number;
  repeatCustomers: number;
  subscriptions: number;
}

export interface CustomerTrends {
  totalCustomers: number;
  newThisMonth: number;
  repeatCustomers: number;
  repeatRate: number;
  activeSubscriptions: number;
  avgSubscriptionDays: number;
  churnedThisMonth: number;
  monthlyTrend: MonthlyCustomerPoint[];
}

export interface FunnelStage {
  stage: string;
  users: number;
  dropOff: number;
  dropOffRate: number;
}

export interface DropOffReason {
  reason: string;
  count: number;
  percentage: number;
}

export interface DropOffAnalysis {
  stages: FunnelStage[];
  topDropOffStage: string;
  dropOffReasons: DropOffReason[];
}

export interface RatingCount {
  stars: number;
  count: number;
}

export interface MoodAnalysis {
  averageRating: number;
  totalReviews: number;
  ratingDistribution: RatingCount[];
  sentimentScore: number;
  topPositiveThemes: string[];
  topNegativeThemes: string[];
}

export interface Customer {
  id: number;
  name: string;
  email: string;
  isRepeat: boolean;
  isSubscribed: boolean;
  /** @nullable */
  subscriptionDays?: number | null;
  /** @nullable */
  subscriptionPlan?: string | null;
  totalOrders?: number;
  totalSpend?: number;
  /** @nullable */
  source?: string | null;
  createdAt: string;
}

export interface CustomerInput {
  name: string;
  email: string;
  isRepeat?: boolean;
  isSubscribed?: boolean;
  /** @nullable */
  subscriptionDays?: number | null;
  /** @nullable */
  subscriptionPlan?: string | null;
  totalOrders?: number;
  totalSpend?: number;
  /** @nullable */
  source?: string | null;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  onSale: boolean;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  description?: string | null;
  isNappySub?: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  category: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  onSale?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  description?: string | null;
  isNappySub?: boolean;
}

export interface Review {
  id: number;
  customerId: number;
  productId: number;
  rating: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  sentiment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  customerId: number;
  productId: number;
  rating: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  sentiment?: string | null;
}

export interface DiagnoseInput {
  /** The stage to focus the diagnosis on (e.g. "checkout", "cart", "banner") */
  funnelStage: string;
  /** @nullable */
  additionalContext?: string | null;
}

export type InsightSeverity =
  (typeof InsightSeverity)[keyof typeof InsightSeverity];

export const InsightSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Insight {
  title: string;
  description: string;
  severity: InsightSeverity;
}

export type ExperimentEffort =
  (typeof ExperimentEffort)[keyof typeof ExperimentEffort];

export const ExperimentEffort = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type ExperimentStatus =
  (typeof ExperimentStatus)[keyof typeof ExperimentStatus];

export const ExperimentStatus = {
  proposed: "proposed",
  active: "active",
  completed: "completed",
  rejected: "rejected",
} as const;

export interface Experiment {
  id: number;
  title: string;
  hypothesis: string;
  expectedImpact: string;
  effort: ExperimentEffort;
  funnelStage: string;
  status?: ExperimentStatus;
  createdAt: string;
}

export interface DiagnosisResult {
  summary: string;
  topInsights: Insight[];
  experiments: Experiment[];
  generatedAt: string;
}

export type ListEventsParams = {
  eventType?: string;
  sessionId?: string;
  limit?: number;
};

export type ListCustomersParams = {
  type?: ListCustomersType;
  limit?: number;
};

export type ListCustomersType =
  (typeof ListCustomersType)[keyof typeof ListCustomersType];

export const ListCustomersType = {
  all: "all",
  repeat: "repeat",
  new: "new",
  subscribed: "subscribed",
} as const;

export type ListProductsParams = {
  onSale?: boolean;
  category?: string;
};

export type ListReviewsParams = {
  productId?: number;
  minRating?: number;
  limit?: number;
};