# lib/api-zod/src/generated/types/customerTrends.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexpoint Funnel IQ — AI-assisted funnel drop-off diagnosis tool and Happy Mom campaign tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyCustomerPoint } from "./monthlyCustomerPoint";

export interface CustomerTrends {
  totalCustomers: number;
  newThisMonth: number;
  repeatCustomers: number;
  repeatRate: number;
  activeSubscriptions: number;
  avgSubscriptionDays: number;
  churnedThisMonth: number;
  monthlyTrend: MonthlyCustomerPoint[];
}