# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexpoint Funnel IQ — AI-assisted funnel drop-off diagnosis tool and Happy Mom campaign tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}